pyproject.toml
[project]
name = "wikipedia-mobile"
version = "0.1.0"
description = ""
authors = [
    {name = "artur.balsis",email = "artur.balsis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "selene (==2.0.0rc9)",
    "requests (==2.32.3)",
    "allure-pytest (==2.14.0)",
    "python-dotenv (==1.1.0)",
    "appium-python-client (==5.0.0)",
    "pydantic (==2.11.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
